Move reused types to types file and clean up component files
